solutions/solution-00/requirements.txt
Flask>=2.2
SQLAlchemy==1.4.20
Flask-SQLAlchemy==3.0.0
flask-cors
flask_jwt_extended
python-dotenv
gunicorn
click>=7.1.2
Jinja2>=3.0
Werkzeug>=2.0
itsdangerous>=2.0
PyJWT<3.0,>=2.0
packaging
